[EGD-7136] Apps setting cleanup and fixes
Apps setting cleanup. Disabled in UI not working options.
