+++ module-apps/application-settings/windows/apps/AlarmClockWindow.cpp
@@ -2,10 +2,9 @@
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #include "AlarmClockWindow.hpp"
-#include "application-settings/ApplicationSettings.hpp"
-#include "application-settings/widgets/SpinBoxOptionSettings.hpp"
-#include "BaseSettingsWindow.hpp"
 
+#include <application-settings/ApplicationSettings.hpp>
+#include <application-settings/widgets/SpinBoxOptionSettings.hpp>
 #include <i18n/i18n.hpp>
 
 namespace gui
